Extract ImportArea, ReviewStep components and unify character state
- Create ImportArea.svelte to deduplicate import textarea+button from
  AbilityRoller and CharacterSheet
- Create ReviewStep.svelte to extract 89-line inline review step from
  App.svelte
- Add makeEmptyCharacter() factory to unify initial state and resetAll(),
  declaring all fields (sex, alignment, age, etc.) upfront
