+++ src/App.svelte
@@ -9,7 +9,7 @@
   import BackstoryEditor from './lib/BackstoryEditor.svelte';
   import CharacterSheet from './lib/CharacterSheet.svelte';
   import CharacterSummary from './lib/CharacterSummary.svelte';
-  import { getCharacterWarnings } from './data/classes.js';
+  import ReviewStep from './lib/ReviewStep.svelte';
   import { initTheme, toggleTheme } from './lib/theme.js';
   import { getCharacterFromUrl, encodeCharacter, decodeCharacter } from './lib/shareCharacter.js';
 
@@ -28,24 +28,35 @@
   let currentStep = $state(0);
   let theme = $state('light');
 
-  // Character data
-  let character = $state({
-    abilities: null,        // Base abilities from rolling
-    rollData: null,         // Original roll data (dice, method, etc.) for reassignment
-    adjustedAbilities: null, // After racial adjustments
-    raceKey: null,
-    race: null,
-    classKey: null,
-    cls: null,
-    levelLimit: null,
-    xpBonus: 0,
-    wizardSchool: null,     // For specialist wizards
-    proficiencies: null,    // Weapon and non-weapon proficiencies
-    equipment: null,        // Purchased equipment
-    spells: null,           // Spellbook/prepared spells
-    name: null,             // Character name
-    backstory: null,        // Character backstory
-  });
+  function makeEmptyCharacter() {
+    return {
+      abilities: null,
+      rollData: null,
+      adjustedAbilities: null,
+      raceKey: null,
+      race: null,
+      classKey: null,
+      cls: null,
+      levelLimit: null,
+      xpBonus: 0,
+      wizardSchool: null,
+      proficiencies: null,
+      equipment: null,
+      spells: null,
+      name: null,
+      backstory: null,
+      sex: null,
+      alignment: null,
+      age: null,
+      height: null,
+      weight: null,
+      eyes: null,
+      hair: null,
+      deity: null,
+    };
+  }
+
+  let character = $state(makeEmptyCharacter());
 
   const SAVE_KEY = 'backdraft-forge-character';
 
@@ -233,23 +244,7 @@
   let showResetConfirm = $state(false);
 
   function resetAll() {
-    character = {
-      abilities: null,
-      rollData: null,
-      adjustedAbilities: null,
-      raceKey: null,
-      race: null,
-      classKey: null,
-      cls: null,
-      levelLimit: null,
-      xpBonus: 0,
-      wizardSchool: null,
-      proficiencies: null,
-      equipment: null,
-      spells: null,
-      name: null,
-      backstory: null,
-    };
+    character = makeEmptyCharacter();
     currentStep = 0;
     showResetConfirm = false;
     hasSavedCharacter = false;
@@ -353,93 +348,7 @@
 
     {:else if currentStep === 3}
       <h2>Review Your Character</h2>
-
-      <div class="character-review">
-        <div class="review-header">
-          <span class="review-race">{character.race.name}</span>
-          <span class="review-class">
-            {#if character.wizardSchool}
-              {character.wizardSchool.name}
-            {:else}
-              {character.cls.name}
-            {/if}
-          </span>
-        </div>
-
-        <div class="review-grid">
-          <div class="review-section">
-            <h4>Ability Scores</h4>
-            <div class="ability-summary">
-              {#each ['STR', 'DEX', 'CON', 'INT', 'WIS', 'CHA'] as ability}
-                <div class="ability-badge">
-                  <span class="ability-label">{ability}</span>
-                  <span class="ability-value">
-                    {character.adjustedAbilities[ability]}{#if ability === 'STR' && character.abilities.exceptionalStr}/{character.abilities.exceptionalStr.toString().padStart(2, '0')}{/if}
-                  </span>
-                </div>
-              {/each}
-            </div>
-          </div>
-
-          <div class="review-section">
-            <h4>Class Info</h4>
-            <div class="info-list">
-              <div class="info-row">
-                <span>Hit Die</span>
-                <span>{character.cls.hitDie}</span>
-              </div>
-              <div class="info-row">
-                <span>Prime Requisite</span>
-                <span>{character.cls.primeRequisite.join(', ')}</span>
-              </div>
-              {#if character.wizardSchool}
-                <div class="info-row">
-                  <span>School</span>
-                  <span>{character.wizardSchool.school}</span>
-                </div>
-                <div class="info-row warning">
-                  <span>Cannot Cast</span>
-                  <span>{character.wizardSchool.oppositionSchools.join(', ')}</span>
-                </div>
-              {/if}
-              {#if character.xpBonus > 0}
-                <div class="info-row highlight">
-                  <span>XP Bonus</span>
-                  <span>+{character.xpBonus}%</span>
-                </div>
-              {/if}
-              {#if character.levelLimit}
-                <div class="info-row warning">
-                  <span>Level Limit</span>
-                  <span>{character.levelLimit}</span>
-                </div>
-              {/if}
-            </div>
-          </div>
-        </div>
-
-        {#if character.cls}
-          {@const reviewWarnings = getCharacterWarnings(character.adjustedAbilities, character.cls, character.classKey)}
-          {#if reviewWarnings.length > 0}
-            <div class="review-warnings">
-              <h4>Advisory</h4>
-              {#each reviewWarnings as warning}
-                <div class="review-warning {warning.severity}">
-                  <span class="warning-icon">{warning.severity === 'concern' ? '⚠' : '△'}</span>
-                  <span>{warning.message}</span>
-                </div>
-              {/each}
-            </div>
-          {/if}
-        {/if}
-
-        <button
-          class="btn-primary"
-          onclick={() => continueToStep(4)}
-        >
-          Continue to Proficiencies
-        </button>
-      </div>
+      <ReviewStep {character} onContinue={() => continueToStep(4)} />
 
     {:else if currentStep === 4}
       <h2>Choose Proficiencies</h2>
@@ -646,159 +555,6 @@
 
   .content {
     min-height: 400px;
-  }
-
-  .ability-summary {
-    display: flex;
-    justify-content: center;
-    gap: 0.75rem;
-    flex-wrap: wrap;
-  }
-
-  .ability-badge {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 0.5rem 0.75rem;
-    background: rgba(201, 162, 39, 0.15);
-    border: 1px solid rgba(201, 162, 39, 0.3);
-    border-radius: 4px;
-    min-width: 50px;
-
-    .ability-label {
-      font-size: 0.7rem;
-      font-weight: 600;
-      color: var(--text-muted);
-      letter-spacing: 0.05em;
-    }
-
-    .ability-value {
-      font-family: 'Cinzel', serif;
-      font-size: 1.25rem;
-      font-weight: 700;
-      color: var(--text-primary);
-    }
-  }
-
-  // Character Review
-  .character-review {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-    align-items: center;
-
-    .btn-primary {
-      margin-top: 1rem;
-    }
-  }
-
-  .review-header {
-    display: flex;
-    justify-content: center;
-    gap: 0.5rem;
-    font-size: 1.5rem;
-    font-family: 'Cinzel', serif;
-    width: 100%;
-
-    .review-race {
-      color: var(--text-muted);
-    }
-
-    .review-class {
-      color: var(--text-primary);
-      font-weight: 600;
-    }
-  }
-
-  .review-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-    gap: 1.5rem;
-    width: 100%;
-  }
-
-  .review-section {
-    padding: 1rem;
-    background: var(--bg-panel);
-    border-radius: 4px;
-
-    h4 {
-      margin: 0 0 1rem;
-      font-size: 1rem;
-      color: var(--text-body);
-      border-bottom: 1px solid var(--border-color);
-      padding-bottom: 0.25rem;
-    }
-  }
-
-  .info-list {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-
-  .info-row {
-    display: flex;
-    justify-content: space-between;
-    font-size: 0.9rem;
-
-    span:first-child {
-      color: var(--text-muted);
-    }
-
-    span:last-child {
-      font-weight: 600;
-      color: var(--text-primary);
-    }
-
-    &.highlight span:last-child {
-      color: #228b22;
-    }
-
-    &.warning span:last-child {
-      color: var(--gold-dark);
-    }
-  }
-
-  // Advisory Warnings
-  .review-warnings {
-    width: 100%;
-    margin-top: 0.5rem;
-
-    h4 {
-      font-size: 0.85rem;
-      text-transform: uppercase;
-      letter-spacing: 0.05em;
-      color: var(--text-muted);
-      margin: 0 0 0.5rem;
-    }
-  }
-
-  .review-warning {
-    display: flex;
-    align-items: flex-start;
-    gap: 0.5rem;
-    padding: 0.4rem 0.6rem;
-    border-radius: 4px;
-    font-size: 0.85rem;
-    line-height: 1.3;
-    margin-bottom: 0.3rem;
-
-    .warning-icon {
-      flex-shrink: 0;
-    }
-
-    &.caution {
-      background: rgba(184, 148, 60, 0.1);
-      color: var(--gold-dark);
-      border-left: 3px solid var(--gold-dark);
-    }
-
-    &.concern {
-      background: rgba(180, 60, 40, 0.1);
-      color: #b43c28;
-      border-left: 3px solid #b43c28;
-    }
   }
 
   /* Print Styles */
